requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT~=2.10.1
ruff==0.11.8
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
